chore(wallets): extract WalletRow into its own component
