+++ extension/src/popup/components/account/WalletRow/styles.scss
@@ -1,0 +1,81 @@
+@use "../../../styles/utils.scss" as *;
+
+.WalletRow {
+  display: flex;
+  width: 100%;
+  margin-bottom: #{pxToRem(24px)};
+
+  &__identicon {
+    width: 50px;
+    height: 50px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    position: relative;
+    cursor: pointer;
+
+    .identicon-wrapper {
+      background: var(--color-white);
+      border-radius: 2rem;
+      height: 2.5rem;
+      width: 2.5rem;
+      margin: 0.125rem;
+      padding: 0.5rem;
+      border-style: solid;
+      border-width: 1.5px;
+      border-color: var(--sds-clr-gray-03);
+
+      &.is-selected {
+        border-width: 2px;
+      }
+    }
+
+    &__selected-check {
+      width: 12px;
+      height: 12px;
+      position: absolute;
+      bottom: 6px;
+      left: 5px;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      padding: 2px;
+      border-radius: 50%;
+    }
+  }
+
+  &__options {
+    width: 50px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+  }
+
+  &__details {
+    flex: 1;
+    min-width: 0;
+    display: flex;
+    flex-direction: column;
+    cursor: pointer;
+    padding-left: #{pxToRem(12px)};
+
+    .detail-name {
+      margin: 0 !important; // sds override
+    }
+
+    .detail-short-key {
+      margin: 0 !important; // sds override
+      line-height: 1rem;
+      color: var(--sds-clr-gray-11) !important; // sds override
+      font-size: #{pxToRem(12px)};
+    }
+
+    .detail-indentifier {
+      margin: 0 !important; // sds override
+      line-height: 1rem;
+      color: var(--sds-clr-gray-11) !important; // sds override
+      font-size: #{pxToRem(12px)};
+    }
+  }
+}
