+++ extension/src/popup/views/Wallets/index.tsx
@@ -3,13 +3,11 @@ import { useTranslation } from "react-i18next";
 import { useDispatch, useSelector } from "react-redux";
 import { Navigate, useNavigate } from "react-router-dom";
 import { Button, CopyText, Icon, Notification } from "@stellar/design-system";
-import classNames from "classnames";
 
 import { AppDispatch } from "popup/App";
 import { ROUTES } from "popup/constants/routes";
 import { SubviewHeader } from "popup/components/SubviewHeader";
 import { View } from "popup/basics/layout/View";
-import { IdenticonImg } from "popup/components/identicons/IdenticonImg";
 import {
   makeAccountActive,
   allAccountsSelector,
@@ -19,9 +17,6 @@ import {
   clearBalancesForAccount,
   clearCollectiblesForAccount,
 } from "popup/ducks/cache";
-import IconEllipsis from "popup/assets/icon-ellipsis.svg";
-import { truncatedPublicKey } from "helpers/stellar";
-import { getColorPubKey } from "helpers/stellarIdenticon";
 import { LoadingBackground } from "popup/basics/LoadingBackground";
 import { useGetWalletsData } from "./hooks/useGetWalletsData";
 import { RequestState } from "constants/request";
@@ -30,89 +25,11 @@ import { AppDataType } from "helpers/hooks/useGetAppData";
 import { newTabHref } from "helpers/urls";
 import { navigateTo, openTab } from "popup/helpers/navigate";
 import { reRouteOnboarding } from "popup/helpers/route";
-import { WalletType } from "@shared/constants/hardwareWallet";
-import { AddWallet } from "popup/components/account/AddWallet";
+import { WalletRow } from "popup/components/account/WalletRow";
 import { RenameWallet } from "popup/components/account/RenameWallet";
+import { AddWallet } from "popup/components/account/AddWallet";
 
 import "./styles.scss";
-
-interface WalletRowProps {
-  isFetchingTokenPrices: boolean;
-  accountName: string;
-  accountValue: string;
-  isImported: boolean;
-  hardwareWalletType?: WalletType;
-  isSelected: boolean;
-  publicKey: string;
-  onClick: (publicKey: string) => unknown;
-  setOptionsOpen: (publicKey: string) => unknown;
-}
-
-const WalletRow = ({
-  isFetchingTokenPrices,
-  accountName,
-  accountValue,
-  isImported,
-  hardwareWalletType,
-  isSelected,
-  publicKey,
-  onClick,
-  setOptionsOpen,
-}: WalletRowProps) => {
-  const shortPublicKey = truncatedPublicKey(publicKey);
-  const identiconWrapperStyles = classNames("identicon-wrapper", {
-    "is-selected": isSelected,
-  });
-  const selectedBorderColorRgb = getColorPubKey(publicKey);
-  const isSelectedColor = `rgb(${selectedBorderColorRgb.r} ${selectedBorderColorRgb.g} ${selectedBorderColorRgb.b} / 100%`;
-  const borderColor = isSelected ? isSelectedColor : "#232323";
-
-  let subTitle = accountValue
-    ? `${shortPublicKey} - ${accountValue}`
-    : shortPublicKey;
-  if (isFetchingTokenPrices && !accountValue) {
-    subTitle = `${shortPublicKey} - ...`;
-  }
-  const { t } = useTranslation();
-  const walletIdentifier =
-    hardwareWalletType || isImported ? t("Imported") : "";
-  return (
-    <div className="WalletRow">
-      <div
-        className="WalletRow__identicon"
-        onClick={() => onClick(publicKey)}
-        data-testid="wallet-row-select"
-      >
-        <div
-          className={identiconWrapperStyles}
-          style={{ borderColor: borderColor }}
-        >
-          <IdenticonImg publicKey={publicKey} />
-        </div>
-        {isSelected ? (
-          <div
-            className="WalletRow__identicon__selected-check"
-            style={{ backgroundColor: isSelectedColor }}
-          >
-            <Icon.Check width="14px" height="14px" />
-          </div>
-        ) : null}
-      </div>
-      <div className="WalletRow__details" onClick={() => onClick(publicKey)}>
-        <p className="detail-name">{accountName}</p>
-        <p className="detail-short-key">{subTitle}</p>
-        <p className="detail-short-key">{walletIdentifier}</p>
-      </div>
-      <div
-        className="WalletRow__options"
-        data-testid="wallet-row-options"
-        onClick={() => setOptionsOpen(publicKey)}
-      >
-        <img src={IconEllipsis} alt={t("wallet action options")} />
-      </div>
-    </div>
-  );
-};
 
 export const Wallets = () => {
   const activeOptionsRef = useRef<HTMLDivElement>(null);
