+++ extension/src/popup/views/Wallets/styles.scss
@@ -1,135 +1,5 @@
 @use "../../styles/utils.scss" as *;
 
-.WalletRow {
-  display: flex;
-  width: 100%;
-  margin-bottom: #{pxToRem(24px)};
-
-  &__identicon {
-    width: 50px;
-    height: 50px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    cursor: pointer;
-
-    .identicon-wrapper {
-      background: var(--color-white);
-      border-radius: 2rem;
-      height: 2.5rem;
-      width: 2.5rem;
-      margin: 0.125rem;
-      padding: 0.5rem;
-      border-style: solid;
-      border-width: 1.5px;
-      border-color: var(--sds-clr-gray-03);
-
-      &.is-selected {
-        border-width: 2px;
-      }
-    }
-
-    &__selected-check {
-      width: 12px;
-      height: 12px;
-      position: absolute;
-      bottom: 6px;
-      left: 5px;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      padding: 2px;
-      border-radius: 50%;
-    }
-  }
-
-  &__options {
-    width: 50px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-  }
-
-  &__options-actions {
-    display: flex;
-    flex-direction: column;
-    position: absolute;
-    background-color: var(--sds-clr-gray-01);
-    border: 1px solid var(--sds-clr-gray-06);
-    border-radius: 6px;
-    margin-top: -2rem;
-    padding: 0.25rem;
-    right: 2rem;
-    width: 10.75rem;
-    z-index: var(--z-index--banner);
-
-    &__row {
-      padding: 0.375rem 0.5rem;
-      width: 100%;
-      border-radius: 6px;
-
-      &:hover {
-        background-color: var(--sds-clr-gray-05);
-      }
-
-      .CopyText__content,
-      .CopyText {
-        width: 100%;
-      }
-
-      .action-copy {
-        width: 100%;
-        display: flex;
-        justify-content: space-between;
-
-        svg {
-          color: var(--sds-clr-gray-09);
-        }
-      }
-    }
-
-    &__label {
-      font-size: 0.875rem;
-      font-weight: var(--font-weight-medium);
-    }
-
-    &__background {
-      height: 100vh;
-      position: absolute;
-      width: 100vw;
-    }
-  }
-
-  &__details {
-    flex: 1;
-    min-width: 0;
-    display: flex;
-    flex-direction: column;
-    cursor: pointer;
-    padding-left: #{pxToRem(12px)};
-
-    .detail-name {
-      margin: 0 !important; // sds override
-    }
-
-    .detail-short-key {
-      margin: 0 !important; // sds override
-      line-height: 1rem;
-      color: var(--sds-clr-gray-11) !important; // sds override
-      font-size: #{pxToRem(12px)};
-    }
-
-    .detail-indentifier {
-      margin: 0 !important; // sds override
-      line-height: 1rem;
-      color: var(--sds-clr-gray-11) !important; // sds override
-      font-size: #{pxToRem(12px)};
-    }
-  }
-}
-
 .RenameWalletWrapper {
   position: absolute;
   width: 100%;
@@ -144,4 +14,54 @@
   flex-direction: column;
   background-color: var(--sds-clr-gray-01);
   z-index: var(--z-index--banner);
+}
+
+.WalletRow__options-actions {
+  display: flex;
+  flex-direction: column;
+  position: absolute;
+  background-color: var(--sds-clr-gray-01);
+  border: 1px solid var(--sds-clr-gray-06);
+  border-radius: 6px;
+  margin-top: -2rem;
+  padding: 0.25rem;
+  right: 2rem;
+  width: 10.75rem;
+  z-index: var(--z-index--banner);
+
+  &__row {
+    padding: 0.375rem 0.5rem;
+    width: 100%;
+    border-radius: 6px;
+
+    &:hover {
+      background-color: var(--sds-clr-gray-05);
+    }
+
+    .CopyText__content,
+    .CopyText {
+      width: 100%;
+    }
+
+    .action-copy {
+      width: 100%;
+      display: flex;
+      justify-content: space-between;
+
+      svg {
+        color: var(--sds-clr-gray-09);
+      }
+    }
+  }
+
+  &__label {
+    font-size: 0.875rem;
+    font-weight: var(--font-weight-medium);
+  }
+
+  &__background {
+    height: 100vh;
+    position: absolute;
+    width: 100vw;
+  }
 }
